pyproject.toml
# Rename to build stable version
# This is stable version
[build-system]
requires = ["setuptools>=82.0.1"]
build-backend = "setuptools.build_meta"

[project]
name = "je_api_testka"
version = "0.0.139"
authors = [
    { name = "JE-Chen", email = "jechenmailman@gmail.com" },
]
description = "Requests Automation Framework"
requires-python = ">=3.10"
license-files = ["LICENSE"]
dependencies = [
    "requests", "Flask", "httpx", "defusedxml"
]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Win32 (MS Windows)",
    "Environment :: MacOS X",
    "Environment :: X11 Applications",
    "Operating System :: OS Independent"
]

[project.urls]
Homepage = "https://github.com/Intergration-Automation-Testing/APITestka"
Documentation = "https://apitestka.readthedocs.io/en/latest/"
Code = "https://github.com/Intergration-Automation-Testing/APITestka"

[project.readme]
file = "README.md"
content-type = "text/markdown"

[tool.setuptools.packages]
find = { namespaces = false }

[project.scripts]
apitestka = "je_api_testka.cli.cli_main:main"
apitestka-mcp = "je_api_testka.mcp_server.server:main"

[project.entry-points."pytest11"]
apitestka = "je_api_testka.pytest_plugin.plugin"

[project.optional-dependencies]
gui = ["PySide6==6.11.0", "qt-material"]
websocket = ["websockets>=12.0"]
schema = ["jsonschema>=4.0", "jsonpath-ng>=1.6"]
security = ["pyjwt>=2.8", "botocore>=1.34"]
otel = ["opentelemetry-api>=1.25", "opentelemetry-sdk>=1.25"]
mcp = ["mcp>=1.0"]

[tool.pytest.ini_options]
testpaths = ["test"]
asyncio_mode = "auto"
timeout = 30

[tool.bandit]
# Test code uses assert by design (pytest); see CLAUDE.md "Test Code Exemptions".
exclude_dirs = ["test", "tests", "build", "dist", ".venv", "venv"]
